[VACGOM-69] refactor: BabyManager Entity 코틀린 엔티티로 변경
