+++ src/main/kotlin/kr/co/vacgom/api/babymanager/repository/BabyManagerJpaRepository.kt
@@ -1,6 +1,6 @@
-package kr.co.vacgom.api.baby.repository
+package kr.co.vacgom.api.babymanager.repository
 
-import kr.co.vacgom.api.babymanager.BabyManager
+import kr.co.vacgom.api.babymanager.domain.BabyManager
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface BabyManagerJpaRepository: JpaRepository<BabyManager, Long>
