+++ src/main/kotlin/kr/co/vacgom/api/babymanager/repository/BabyManagerRepository.kt
@@ -1,6 +1,7 @@
-package kr.co.vacgom.api.baby.repository
+package kr.co.vacgom.api.babymanager.repository
 
-import kr.co.vacgom.api.babymanager.BabyManager
+import kr.co.vacgom.api.babymanager.domain.BabyManager
+
 
 interface BabyManagerRepository {
     fun save(babyManager: BabyManager): BabyManager
